requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
colour-science=0.4.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
imageio=2.33.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
lz4-c=1.9.4=h6a678d5_0
mkl=2023.1.0=h213fc3f_46344
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.4=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.12=h7f8727e_0
pillow=9.3.0=py38hace64e9_1
pip=23.3.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
pytorch=1.9.1=py3.8_cuda10.2_cudnn7.6.5_0
readline=8.2=h5eee18b_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchvision=0.10.1=py38_cu102
tqdm=4.66.1=pypi_0
trilinear=0.0.0=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
